requirements.txt
python-telegram-bot[webhooks]==13.15
quart>=0.18.4
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
requests>=2.31.0 